Moved address object DTO to wizard service for reuse
