+++ responders/src/UI/embc-responder/src/app/feature-components/wizard/wizard.service.ts
@@ -1,9 +1,12 @@
 import { Injectable } from '@angular/core';
+import { Address } from 'src/app/core/api/models';
+import { AddressModel } from 'src/app/core/models/address.model';
 import {
   WizardSidenavModel,
   WizardSidenavModelValues
 } from 'src/app/core/models/wizard-sidenav.model';
 import { CacheService } from 'src/app/core/services/cache.service';
+import { Community } from 'src/app/core/services/locations.service';
 
 @Injectable({ providedIn: 'root' })
 export class WizardService {
@@ -61,5 +64,61 @@ export class WizardService {
       }
       return menu;
     });
+  }
+
+  /**
+   * Map an address from the wizard to an address usable by the API
+   *
+   * @param addressObject An Address as defined by the site's address forms
+   * @returns Address object as defined by the API
+   */
+  public setAddressObjectForDTO(addressObject: AddressModel): Address {
+    const address: Address = {
+      addressLine1: addressObject.addressLine1,
+      addressLine2: addressObject.addressLine2,
+      countryCode: addressObject.country.code,
+      communityCode:
+        (addressObject.community as Community).code === undefined
+          ? null
+          : (addressObject.community as Community).code,
+      city:
+        (addressObject.community as Community).code === undefined &&
+        typeof addressObject.community === 'string'
+          ? addressObject.community
+          : null,
+      postalCode: addressObject.postalCode,
+      stateProvinceCode:
+        addressObject.stateProvince === null
+          ? null
+          : addressObject.stateProvince.code
+    };
+
+    return address;
+  }
+
+  /**
+   * Map an address from the API to an address usable by the wizard form
+   *
+   * @param addressObject Address object as defined by the API
+   * @returns An Address as defined by the site's address forms
+   */
+  public setAddressObjectForForm(addressObject: AddressModel): AddressModel {
+    const address: AddressModel = {
+      addressLine1: addressObject.addressLine1,
+      addressLine2: addressObject.addressLine2,
+      communityCode: addressObject.communityCode,
+      countryCode: addressObject.countryCode,
+      stateProvinceCode: addressObject.stateProvinceCode,
+      country: addressObject.country,
+      postalCode: addressObject.postalCode,
+      stateProvince: addressObject.stateProvince,
+      community:
+        addressObject.city !== null
+          ? addressObject.city
+          : addressObject.community,
+      city: null
+    };
+
+    return address;
   }
 }
